requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-chroma
chromadb
pypdf
python-dotenv
sqlalchemy 
psycopg2
langchain-groq
langsmith